+++ gcc/gcc/testsuite/g++.dg/parse/cast1.C
@@ -1,10 +1,11 @@
 // PR c++/13536
+// { dg-require-effective-target ilp32 } */
 // { dg-options "-w" }
 
 #include <typeinfo>
 
 void f() {
-  (long((char*)0));
-  sizeof ((long((char*)0)));
-  typeid ((long((char*)0)));
+  (int((char*)0));
+  sizeof ((int((char*)0)));
+  typeid ((int((char*)0)));
 }
